backend/requirements.txt
Django==4.0.1
djangorestframework==3.13.1
django-cors-headers==3.11.0
psycopg2-binary==2.9.11
python-decouple==3.8
gunicorn==20.1.0
whitenoise==6.2.0
